Add builtin.breakpoint
